rag/requirements.txt
fastapi
pinecone-client
langchain
sentence-transformers
uvicorn[standard]